requirements.txt
langchain
python-dotenv
ipykernel
langchain-community
langchain-google-genai
langchain_community 
faiss-cpu
pypdf
streamlit
google-generativeai 
langchain_google_genai 
